move playwright to root
