+++ .gitignore
@@ -8,3 +8,7 @@
 **/.DS_Store
 
 .turbo
+
+/test-results/
+/playwright-report/
+/playwright/.cache/
