+++ apps/web/package.json
@@ -4,9 +4,7 @@
   "private": true,
   "scripts": {
     "build": "cross-env HOST_NAME=https://bypass-links.vercel.app next build",
-    "dev": "cross-env HOST_NAME=http://localhost:3000 next dev",
-    "e2e": "playwright test",
-    "e2e:report": "playwright show-report"
+    "dev": "cross-env HOST_NAME=http://localhost:3000 next dev"
   },
   "dependencies": {
     "@bypass/shared": "file:packages/shared",
@@ -29,13 +27,11 @@
   },
   "devDependencies": {
     "@bypass/configs": "file:packages/configs",
-    "@playwright/test": "1.29.1",
     "@svgr/webpack": "6.5.1",
     "@types/md5": "2.3.2",
     "@types/react-window": "1.8.5",
     "@typescript-eslint/eslint-plugin": "5.47.0",
     "cross-env": "7.0.3",
-    "dotenv": "16.0.3",
     "eslint": "8.30.0",
     "eslint-config-next": "13.1.0",
     "eslint-config-prettier": "8.5.0",
@@ -45,7 +41,6 @@
     "eslint-webpack-plugin": "3.2.0",
     "fork-ts-checker-webpack-plugin": "7.2.14",
     "next-pwa": "5.6.0",
-    "playwright": "1.29.1",
     "prettier": "2.8.1",
     "typescript": "4.9.4"
   }
